+++ tests/services/test_alerts.py
@@ -416,3 +416,15 @@ class TestAlertService(object):
             and post_data["groups"][0]["filters"][0]["term"] == "state"
             and post_data["groups"][0]["filters"][0]["value"] == "OPEN"
         )
+
+    def test_get_aggregate_data_calls_post_with_expected_url_and_data(
+        self, mock_connection, user_context
+    ):
+        alert_service = AlertService(mock_connection, user_context)
+        alert_service.get_aggregate_data("alert-id")
+        assert (
+            mock_connection.post.call_args[0][0]
+            == "/svc/api/v1/query-details-aggregate"
+        )
+        post_data = mock_connection.post.call_args[1]["json"]
+        assert post_data["alertId"] == "alert-id"